fix config (#121)
* add time zone

* Merge branch 'master' of https://github.com/NemTechCN/NEM2_Desktop_Wallet_CN

# Conflicts:
#	src/core/utils/utils.ts
#	src/language/en-US.ts
#	src/language/zh-CN.ts
#	src/main.ts

* Merge branch 'master' of https://github.com/NemTechCN/NEM2_Desktop_Wallet_CN

# Conflicts:
#	src/core/utils/utils.ts
#	src/language/en-US.ts
#	src/language/zh-CN.ts
#	src/main.ts

* Merge branch 'master' of https://github.com/NemTechCN/NEM2_Desktop_Wallet_CN

# Conflicts:
#	src/core/utils/utils.ts
#	src/language/en-US.ts
#	src/language/zh-CN.ts
#	src/main.ts

* update complete transaction detail

* update promise to rxjs

* update promise to rxjs

* Merge branch 'master' of https://github.com/nemfoundation/nem2-desktop-wallet into nemfoundation-master

# Conflicts:
#	src/common/css/ivewWindows.less
#	src/common/css/iviewMac.less
#	src/common/vue/collection-record/CollectionRecordTs.ts
#	src/common/vue/menu-bar/MenuBarTs.ts
#	src/config/index.ts
#	src/core/api/transactionApi.ts
#	src/core/utils/format.ts
#	src/core/utils/wallet.ts
#	src/language/en-US.ts
#	src/language/zh-CN.ts
#	src/views/monitor/monitor-dashboard/MonitorDashBoardTs.ts

* Merge branch 'master' of https://github.com/nemfoundation/nem2-desktop-wallet

# Conflicts:
#	src/language/zh-CN.ts
#	src/views/monitor/monitor-transfer/transactions/transfer-transaction/TransferTransactionTs.ts

* Merge branch 'master' of https://github.com/nemfoundation/nem2-desktop-wallet

# Conflicts:
#	src/language/zh-CN.ts
#	src/views/monitor/monitor-transfer/transactions/transfer-transaction/TransferTransactionTs.ts

* update signature

* Merge branch 'master' of https://github.com/nemfoundation/nem2-desktop-wallet

# Conflicts:
#	src/App.vue
#	src/common/vue/check-password-dialog/CheckPasswordDialogTs.ts
#	src/common/vue/menu-bar/MenuBarTs.ts
#	src/config/index.ts
#	src/core/api/AccountApiRxjs.ts
#	src/core/api/BlockApiRxjs.ts
#	src/core/api/ListenerApiRxjs.ts
#	src/core/api/MosaicApiRxjs.ts
#	src/core/api/MultisigApiRxjs.ts
#	src/core/api/NamespaceApiRxjs.ts
#	src/core/api/TransactionApiRxjs.ts
#	src/core/api/WalletApiRxjs.ts
#	src/core/utils/wallet.ts
#	src/views/monitor/monitor-dashboard/MonitorDashBoard.vue
#	src/views/monitor/monitor-transfer/transactions/multisig-transfer-transaction/MultisigTransferTransactionTs.ts
#	src/views/monitor/monitor-transfer/transactions/transfer-transaction/TransferTransactionTs.ts
#	src/views/service/mosaic/mosaic-function/mosaic-transaction/MosaicTransactionTs.ts
#	src/views/service/multisig/multisig-functions/multisig-conversion/MultisigConversionTs.ts
#	src/views/service/multisig/multisig-functions/multisig-management/MultisigManagementTs.ts
#	src/views/service/namespace/namespace-function/root-namespace/RootNamespaceTs.ts
#	src/views/service/namespace/namespace-function/sub-namespace/SubNamespaceTs.ts

* update remote sign/announce

* update mosaic and namespace sign/announce

* update mosaic and namespace sign/announce

* add filter

* complete multisig subnamespace transaction
Merge branch 'master' of https://github.com/nemfoundation/nem2-desktop-wallet

# Conflicts:
#	src/core/utils/wallet.ts
#	src/views/service/mosaic/mosaic-function/mosaic-list/mosaic-unAlias-dialog/MosaicUnAliasDialogTs.ts

* fix multisig mosaic creation

* fix multisig mosaic creation

* getMosaicList

* fix namespaceList duration

* udpate stylesheet

* update inner transaction detail in dashboard

* Merge branch 'master' of https://github.com/NemTechCN/nem2-desktop-wallet into task_42/getMosaicList

# Conflicts:
#	src/core/utils/wallet.ts

* fix bugs

* Merge branch 'task_42/getMosaicList' of https://github.com/NemTechCN/nem2-desktop-wallet

# Conflicts:
#	src/views/monitor/monitor-panel/MonitorPanelTs.ts

* fix wallet.ts

* update offline

* update offline

* update offline

* del not used import

* del not used import

* del not used import

* export networkType

* fix infimation locale

* export networkType

* del not used import

* update  timezone

* update  timezone

* update file name

* Enable message encryption

* Enable message encryption

* Enable message encryption

* Enable message encryption

* fix bug package.ts

* fix update wallet password

* fix update wallet password

* update config/index.ts

* fix config

* update mosaic alias namespace

* fix config

* update mosaic alias namespace

* update mosaic alias namespace

* update vee-validator

* format code

* update vee-validate

* upadte mapstate and config

* split configur param

* update mapstate and configure

* update mapstate and configure;fix time bug

* update mapstate and configure;fix time bug

* fix mosaic list bug

* fix qrcode

* update build script

* update README.md

* fix qrcode

* update animation

* fix mosaicAliasMap

* add hint in setting lock-password

* fix dash-board transaction list and balance

* fix dash-board transaction list and balance

* fix address alias

* fix address alias

* update toLocalString

* fix relative mosaic amount

* fix relative mosaic amount

* Revert "fix relative mosaic amount"

This reverts commit eb227e29

* Revert "fix relative mosaic amount"

This reverts commit eb227e29

* update relative mosaic amount

* update relative mosaic amount

* fix wallet balance

* update mosaic amount (to be relative)

* Update UI(delete wallet)

* Update UI(delete wallet)

* update dashboad transaction record

* fix root-namespace check bug

* add unalias in namespace list

* update namespace list in namespace panel

* Merge branch 'master' of https://github.com/nemfoundation/nem2-desktop-wallet

# Conflicts:
#	src/common/vue/menu-bar/MenuBarTs.ts
#	src/config/formData.ts
#	src/config/index.ts
#	src/views/community/vote/VoteTs.ts
#	src/views/monitor/monitor-dashboard/MonitorDashBoardTs.ts
#	src/views/monitor/monitor-panel/MonitorPanelTs.ts
#	src/views/monitor/monitor-transfer/transactions/transfer-transaction/TransferTransactionTs.ts
#	src/views/service/mosaic/mosaic-function/mosaic-list/MosaicListTs.ts

* update stylesheet

* update lineChart

* update lineChart

* fix bugs

* fix bugs

* delete gas

* solve conflects

* update vee-validate

* fix config

* fix receipt

* fix config

* Resolving conflicts
