+++ src/common/vue/collection-record/CollectionRecordTs.ts
@@ -5,7 +5,7 @@ import {
     getCurrentMonthFirst, getCurrentMonthLast, formatNumber,
     renderMosaics, renderMosaicNames, renderMosaicAmount
 } from '@/core/utils'
-import {TransferType} from '@/config/index.ts'
+import {transferType} from "@/config/types"
 
 @Component({
     computed: {...mapState({activeAccount: 'account', app: 'app'})},
@@ -20,11 +20,11 @@ export class CollectionRecordTs extends Vue {
     currentMonthFirst: number = 0
     currentMonth: string = ''
     transactionDetails: any = []
-    TransferType = TransferType
+    transferType = transferType
     renderMosaics = renderMosaics
     renderMosaicNames = renderMosaicNames
     renderMosaicAmount = renderMosaicAmount
-    
+
     @Prop({
         default: () => {
             return 0
@@ -56,11 +56,11 @@ export class CollectionRecordTs extends Vue {
                 && item.txHeader.date <= currentMonthLast && item.txHeader.date >= currentMonthFirst))
         if (!filteredByDate.length) return []
 
-        return this.transactionType === TransferType.SENT
+        return this.transactionType === transferType.SENT
         ? filteredByDate.filter(({txHeader}) => txHeader.tag === 'payment')
         : filteredByDate.filter(({txHeader}) => txHeader.tag !== 'payment')
     }
-    
+
     get mosaicList() {
         return this.activeAccount.mosaics
     }
