+++ src/common/vue/menu-bar/MenuBarTs.ts
@@ -1,5 +1,5 @@
 import routers from '@/router/routers.ts'
-import {Message, isWindows, languageList, localesMap, nodeList} from "@/config/index.ts"
+import {Message, isWindows} from "@/config/index.ts"
 import {AppWallet} from '@/core/utils/wallet'
 import {BlockApiRxjs} from '@/core/api/BlockApiRxjs.ts'
 import monitorSeleted from '@/common/img/window/windowSelected.png'
@@ -8,7 +8,8 @@ import {localSave} from "@/core/utils/utils.ts"
 import {Component, Vue} from 'vue-property-decorator'
 import {windowSizeChange, minWindow, maxWindow, closeWindow} from '@/core/utils/electron.ts'
 import {mapState} from 'vuex'
-import {NamespaceApiRxjs} from "@/core/api/NamespaceApiRxjs"
+import {languageList, nodeList} from "@/config/view";
+import {languageType} from "@/config/types";
 
 @Component({
     computed: {
@@ -33,7 +34,7 @@ export class MenuBarTs extends Vue {
     accountAddress = ''
     txStatusListener = null
     languageList = languageList
-    localesMap = localesMap
+    localesMap = languageType
 
     get isNodeHealthy() {
         return this.app.isNodeHealthy
