+++ src/config/types.ts
@@ -1,0 +1,21 @@
+//language
+export const languageType:any = {
+    'zh-CN': '中文',
+    'en-US': 'English'
+}
+//transaction
+export const transferType = {
+    'RECEIVED': 1,
+    'SENT': 0
+}
+//vote
+export const voteType = {
+    RADIO: 0,
+    MULTIPLE: 1
+}
+//alias
+export const aliasType = {
+    noAlias: 0,
+    mosaicAlias: 1,
+    addressAlias: 2
+}
