+++ src/views/community/vote/PieChart.vue
@@ -7,7 +7,7 @@
 <script lang="ts">
     import echarts from 'echarts'
     import {Component, Vue, Watch, Prop} from 'vue-property-decorator'
-    import {echartsConfigure} from '@/config/index.ts'
+    import {echarts as echartsConfigure} from '@/config/echarts.ts'
 
     @Component
     export default class PieChart extends Vue {
