+++ src/views/community/vote/VoteTs.ts
@@ -1,10 +1,12 @@
 import PieChart from './PieChart.vue'
 import {Component, Vue, Watch} from 'vue-property-decorator'
 import CheckPWDialog from '@/common/vue/check-password-dialog/CheckPasswordDialog.vue'
-import {voteFilterList, alphabet, voteSelectionList, voteActionList, voteType, Message} from '@/config/index.ts'
+import {alphabet, Message} from '@/config/index.ts'
 import {vote} from '@/core/api/logicApi.ts'
 import {mapState} from "vuex"
 import {formatDate} from '@/core/utils/utils.ts'
+import {voteActionList, voteFilterList, voteSelectionList} from "@/config/view";
+import {voteType} from "@/config/types";
 
 @Component({
         components: {
