+++ src/views/monitor/monitor-panel/MonitorPanelTs.ts
@@ -1,10 +1,10 @@
 import {Message} from "@/config/index.ts"
-import {Component, Vue, Watch} from 'vue-property-decorator'
+import {Component, Vue} from 'vue-property-decorator'
 import monitorSeleted from '@/common/img/monitor/monitorSeleted.png'
 import monitorUnselected from '@/common/img/monitor/monitorUnselected.png'
 import {copyTxt, formatXEMamount, formatNumber, localRead, localSave} from '@/core/utils/utils.ts'
 import {mapState} from "vuex"
-import {minitorPanelNavigatorList} from '@/config/index.ts'
+import {monitorPanelNavigatorList} from '@/config/view'
 
 @Component({
     computed: {
@@ -26,7 +26,7 @@ export class MonitorPanelTs extends Vue {
     ischecked = true
     monitorSeleted = monitorSeleted
     monitorUnselected = monitorUnselected
-    navigatorList: any = minitorPanelNavigatorList
+    navigatorList: any = monitorPanelNavigatorList
 
     get xemUsdPrice() {
         return this.app.xemUsdPrice
