+++ src/views/monitor/monitor-remote/MonitorRemoteTs.ts
@@ -1,10 +1,11 @@
-import {Message, formData} from "@/config/index.ts"
+import {Message} from "@/config/index.ts"
 import {Component, Vue} from 'vue-property-decorator'
 import {AccountLinkTransaction, UInt64, LinkAction, Deadline, Password} from "nem2-sdk"
 import {AppWallet} from "@/core/utils/wallet.ts"
 import {AccountApiRxjs} from "@/core/api/AccountApiRxjs.ts"
 import {mapState} from "vuex"
 import {getAbsoluteMosaicAmount} from '@/core/utils'
+import {formData} from "@/config/formDto";
 
 @Component({
     computed: {
