+++ src/views/service/apostille/ApostilleTs.ts
@@ -2,7 +2,8 @@ import {Component, Vue} from 'vue-property-decorator'
 import ApostilleAudit from './apostille-function/apostille-audit/ApostilleAudit.vue'
 import ApostilleCreate from './apostille-function/apostille-create/ApostilleCreate.vue'
 import ApostilleHistory from './apostille-function/apostille-history/ApostilleHistory.vue'
-import {apostilleButtonList} from '@/config/index.ts'
+import {apostilleButtonList} from "@/config/view";
+
 @Component({
     components: {
         ApostilleCreate,
