+++ src/views/service/mosaic/MosaicTs.ts
@@ -1,7 +1,7 @@
 import {Component, Vue} from 'vue-property-decorator'
 import MosaicList from './mosaic-function/mosaic-list/MosaicList.vue'
 import MosaicTransaction from './mosaic-function/mosaic-transaction/MosaicTransaction.vue'
-import {mosaicButtonList} from '@/config/index.ts'
+import {mosaicButtonList} from "@/config/view";
 
 @Component({
     components: {
