+++ src/views/service/mosaic/mosaic-function/mosaic-list/mosaic-alias-dialog/MosaicAliasDialogTs.ts
@@ -1,10 +1,11 @@
 import {AliasActionType, NamespaceId, MosaicId, Password} from "nem2-sdk"
 import {mapState} from "vuex"
-import {Message, namespaceGracePeriodDuration,formData} from "@/config/index.ts"
+import {Message,networkConfig} from "@/config/index.ts"
 import {NamespaceApiRxjs} from "@/core/api/NamespaceApiRxjs.ts"
 import {Component, Vue, Prop, Watch} from 'vue-property-decorator'
 import {EmptyAlias} from "nem2-sdk/dist/src/model/namespace/EmptyAlias"
 import {getAbsoluteMosaicAmount, AppWallet} from '@/core/utils'
+import {formData} from "@/config/formDto";
 
 @Component({
     computed: {
@@ -20,7 +21,7 @@ export class MosaicAliasDialogTs extends Vue {
     show = false
     isCompleteForm = false
     mosaic: any = formData.mosaicAliasForm
-    namespaceGracePeriodDuration = namespaceGracePeriodDuration
+    namespaceGracePeriodDuration = networkConfig.namespaceGracePeriodDuration
 
     @Prop()
     showMosaicAliasDialog: boolean
