+++ src/views/service/mosaic/mosaic-function/mosaic-list/mosaic-edit-dialog/MosaicEditDialogTs.ts
@@ -1,9 +1,10 @@
 import {Component, Vue, Prop, Watch} from 'vue-property-decorator'
 import {mapState} from "vuex"
 import {Password} from 'nem2-sdk'
-import {Message, formData, xemTotalSupply} from "@/config/index.ts"
+import {Message, networkConfig} from "@/config/index.ts"
 import {MosaicApiRxjs} from "@/core/api/MosaicApiRxjs.ts"
 import {getAbsoluteMosaicAmount, AppWallet} from '@/core/utils'
+import {formData} from "@/config/formDto";
 
 @Component({
     computed: {
@@ -16,7 +17,7 @@ export class MosaicEditDialogTs extends Vue {
     activeAccount: any
     isCompleteForm = false
     changedSupply = 0
-    totalSupply = xemTotalSupply
+    totalSupply = networkConfig.maxMosaicAtomicUnits
     mosaic: any = formData.mosaicEditForm
 
     @Prop()
