+++ src/views/service/multisig/MultisigTs.ts
@@ -3,7 +3,8 @@ import MultisigMap from './multisig-functions/multisig-map/MultisigMap.vue'
 import MultisigCosign from './multisig-functions/multisig-cosign/MultisigCosign.vue'
 import MultisigConversion from './multisig-functions/multisig-conversion/MultisigConversion.vue'
 import MultisigManagement from './multisig-functions/multisig-management/MultisigManagement.vue'
-import {multisigButtonList} from '@/config/index.ts'
+import {multisigButtonList} from "@/config/view";
+
 @Component({
     components: {
         MultisigMap,
