+++ src/views/service/multisig/multisig-functions/multisig-map/TopographicMap.vue
@@ -11,7 +11,8 @@
 <script lang="ts">
     import echarts from 'echarts'
     import {mapState} from 'vuex'
-    import {Message, echartsConfigure} from "@/config/index.ts"
+    import {Message} from "@/config/index.ts"
+    import {echarts as  echartsConfigure} from "@/config/echarts.ts"
     import {copyTxt} from '@/core/utils/utils.ts'
     import {Component, Vue} from 'vue-property-decorator'
     import multisignSelfIcon from '@/common/img/service/multisig/multisignSelfIcon.png'
