+++ src/views/wallet/wallet-details/wallet-function/wallet-filter/WalletFilterTs.ts
@@ -1,6 +1,7 @@
 import {Component, Vue} from 'vue-property-decorator'
 import {RestrictionApiRxjs} from '@/core/api/RestrictionApiRxjs.ts'
-import {Message, entityTypeList} from "@/config/index.ts"
+import {Message} from "@/config/index.ts"
+import {transactionTypeList} from "@/config/view"
 import { AppWallet} from "@/core/utils/wallet"
 import {mapState} from "vuex"
 import {
@@ -19,7 +20,7 @@ import {
     computed: {...mapState({activeAccount: 'account'})},
 })
 export class WalletFilterTs extends Vue {
-    entityTypeList = entityTypeList
+    entityTypeList = transactionTypeList
     aliasList = []
     activeAccount: any
     isShowDialog = false
@@ -83,8 +84,8 @@ export class WalletFilterTs extends Vue {
         }
         if (this.filterTypeList[2]) {
             this.formItem.filterList.unshift({
-                label: this.$t(this.currentFilter) + '(' + entityTypeList[this.currentFilter].value + ')',
-                value: entityTypeList[this.currentFilter].value,
+                label: this.$t(this.currentFilter) + '(' + transactionTypeList[this.currentFilter].value + ')',
+                value: transactionTypeList[this.currentFilter].value,
                 modificationType
             })
             return
